requirements.txt
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy[asyncio]==2.0.25
aiosqlite==0.19.0  # SQLite async driver
asyncpg==0.29.0  # PostgreSQL async driver
alembic==1.13.1  # Database migrations

# Web Scraping
beautifulsoup4==4.12.2
lxml==5.1.0
requests==2.31.0
playwright==1.40.0
nest-asyncio==1.6.0

# Background Jobs
apscheduler==3.10.4

# Templates & Frontend
jinja2==3.1.3

# Environment
python-dotenv==1.0.0

# Utilities
python-multipart==0.0.6  # For form data
email-validator==2.2.0

# Production
gunicorn==21.2.0  # Optional: for production with multiple workers
